Add files via upload
The description of the "KnownUnknowns" NFT collection was prepared as a single W3C Turtle RDF file with prefixes (rdf, rdfs, rico, skos, and the prefix for introducing a blank node), a description of the collection of NFTs at the record set level, a description of the minted NFT as a record, its ten SFTs as instances, its analogue original, and this was done for each of the eleven NFTs, their SFTs, and analogue originals. Each NFT is a copy of an analogue photograph, which does not have a URI, so blank nodes had to be introduced to include them in the modelling. However, the collection of NFTs and the NFTs themselves are the focus of the description, so they were modelled as record sets and records, not instances or copies (NFTs), or copies of the copies (SFTs). Each NFT was minted in 10 copies or SFTs (item supply) that do not have their own identifiers and share metadata set at the NFT level. This is the reason for such modelling (and this was discussed with members of the Records in Contexts users’ Google group).